chatbot/requirements.txt
# Strands Agents SDK
strands-agents>=0.1.0

# AWS SDK for Bedrock
boto3>=1.34.0

# HTTP Server
flask>=3.0.0
flask-cors>=4.0.0

# Environment Configuration
python-dotenv>=1.0.0

# HTTP Client for Backend API
requests>=2.31.0

# Testing
hypothesis>=6.92.0
pytest>=7.4.0
pytest-mock>=3.12.0
